chore: reorganise nework domain to distinguish between services
Fixes some misplaced/misnamed interfaces and functions, properly
differentiating between the the migration and regular domain
functionality.
